Tests for plansys2_terminal
