+++ plansys2_terminal/test/CMakeLists.txt
@@ -1,0 +1,8 @@
+ament_add_gtest(terminal_test terminal_test.cpp)
+ament_target_dependencies(terminal_test ${dependencies})
+target_link_libraries(terminal_test terminal readline)
+
+install(DIRECTORY
+  pddl
+  DESTINATION share/${PROJECT_NAME}
+)
